+++ src/test/java/ru/netology/JavaHomeWork7/service/StatsServiceTest.java
@@ -11,7 +11,7 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedSum = 180;
-        int actual = service.sumSales(sales);
+        int actual = (int) service.sumSales(sales);
         Assertions.assertEquals(expectedSum, actual);
     }
 
@@ -20,7 +20,7 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
         long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expectedAverage = 15;
-        long actual = service.averageNumber(sales);
+        int actual = service.averageNumber(sales);
         Assertions.assertEquals(expectedAverage, actual);
     }
